expcap v0.1.1 ExPcap.PacketHeader

This module represents a pcap packet header.

Types

t()

t() :: %ExPcap.PacketHeader{
  incl_len: non_neg_integer(),
  orig_len: non_neg_integer(),
  ts_sec: non_neg_integer(),
  ts_usec: non_neg_integer()
}

Functions

from_file(f, global_header)

from_file(IO.device(), ExPcap.GlobalHeader.t()) :: ExPcap.PacketHeader.t()

Reads a pcap packet header from the file.

read_forward(data)

read_forward(binary()) :: ExPcap.PacketHeader.t()

Reads a packet header from the binary passed in.

read_reversed(data)

read_reversed(binary()) :: ExPcap.PacketHeader.t()

Reads a packet header from the binary passed in, in rerverse byte order.