expcap v0.1.1 Protocol.Dns.Header

A parsed DNS packet header

Types

t()

t() :: %Protocol.Dns.Header{
  aa: bitstring(),
  ancnt: non_neg_integer(),
  arcnt: non_neg_integer(),
  id: binary(),
  nscnt: non_neg_integer(),
  opcode: bitstring(),
  qdcnt: non_neg_integer(),
  qr: bitstring(),
  ra: bitstring(),
  rcode: bitstring(),
  rd: bitstring(),
  tc: bitstring(),
  z: bitstring()
}

Functions

aa_name(aa)

aa_name(binary()) :: :NOT_AUTHORITATIVE | :AUTHORITATIVE | :""

Is this response authoritative?

opcode_name(opcode)

opcode_name(non_neg_integer()) :: :QUERY | :STATUS | :NOTIFY | :UPDATE | :""

What is the op code of this DNS packet?

qr_name(qr)

qr_name(binary()) :: :QUERY | :ANSWER

Is this a query or a response?

ra_name(ra)

ra_name(binary()) :: :NO_RECURSION_AVAILABLE | :RECURSION_AVAILABLE | :""

Is recursion available?

rcode_name(rcode)

rcode_name(non_neg_integer()) :: :atom

What is the r code of this DNS packet?

rd_name(rd)

rd_name(binary()) :: :NO_RECURSION_DESIRED | :RECURSION_DESIRED | :""

Is recursion desired?

tc_name(tc)

tc_name(binary()) :: :NOT_TRUNCATED | :TRUNCATED | :""

Is this response truncated?

z_name(z)

z_name(binary()) :: :atom

The first bit is reserved. The second bit indciates if the response was authenticated or not. The third bit indciates if the data was authenticated or not.