ExPcap.MagicNumber (expcap v0.1.2)

This module represents a 'magic number' from a pcap header. The magic number not only contains a known value, but the value indicates the order in which bytes should be read AND whether or not datetimes use milliseconds or nanoseconds.

Types

t()

Specs

t() :: %ExPcap.MagicNumber{
  magic: non_neg_integer(),
  nanos: boolean(),
  reverse_bytes: boolean()
}

Functions

bytes_in_magic()

Specs

bytes_in_magic() :: non_neg_integer()

Returns the number of bytes contained in the magic number.

from_file(f)

Specs

from_file(IO.device()) :: t()

Reads the magic number from the file passed in.

magic_number(int1, int2, int3, int4)

Specs

magic_number(212, 195, 178, 161) :: t()

magic_number(161, 178, 195, 212) :: t()

magic_number(161, 178, 60, 77) :: t()

magic_number(77, 60, 178, 161) :: t()

Returns a magic number that indicates that the bytes need to be reversed when read and that datetimes are in nanoseconds.

read_magic(data)

Specs

read_magic(binary()) :: t()

This reads the bytes of the magic number and matches them with the appropriate interpretation of the magic number.